enable graphviz output and reverted counter to use long.
git-svn-id: https://gnocchi.svn.sourceforge.net/svnroot/gnocchi/trunk@50 270d8c9f-4531-0410-b52b-96252cf29c44

+++ analyser.hpp
@@ -42,7 +42,7 @@ public:
 	void calculate_npath_2(const Graph&, data_model&);
 	void report(int /*npath_threshold*/);
 
-    typedef boost::signal<void (const gcov_reader&, data_model&, std::vector<long long>)>  line_insert_signal_type;
+    typedef boost::signal<void (const gcov_reader&, data_model&, std::vector<long>)>  line_insert_signal_type;
 	line_insert_signal_type on_complexity_calculated;
 private:
 	reporter& reporter_;
